Initial take of Linux/NuttX exclusive event loop
This will likely undergo quite a few more changes, as I am considering
changing to rb3ptr for the red/black tree implementation. The goal here
is to keep the code footprint to a minimum and minimize any duplication
between libtuv and mininode. Another consideration is overall amortized
system call overhead on Linux and code reuse on NuttX.